backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
bcrypt==4.1.2
reportlab==4.0.7
jinja2==3.1.2
requests==2.31.0
schedule==1.2.1
